+++ src/main/java/nl/dtls/fairdatapoint/api/validator/DurationValidator.java
@@ -20,37 +20,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package nl.dtls.fairdatapoint.service.index.event;
+package nl.dtls.fairdatapoint.api.validator;
 
+import javax.validation.ConstraintValidator;
+import javax.validation.ConstraintValidatorContext;
 
-import nl.dtls.fairdatapoint.api.dto.index.event.EventDTO;
-import nl.dtls.fairdatapoint.entity.index.event.AdminTrigger;
-import nl.dtls.fairdatapoint.entity.index.event.Event;
-import org.springframework.security.core.Authentication;
-import org.springframework.stereotype.Service;
+import java.time.Duration;
 
-import javax.servlet.http.HttpServletRequest;
+import static nl.dtls.fairdatapoint.util.ValueFactoryHelper.i;
 
-@Service
-public class EventMapper {
+public class DurationValidator implements ConstraintValidator<ValidDuration, String> {
 
-    private static final Integer VERSION = 1;
-
-    public EventDTO toDTO(Event event) {
-        return new EventDTO(
-                event.getUuid(),
-                event.getType(),
-                event.getCreated().toString(),
-                event.getFinished().toString()
-        );
+    @Override
+    public void initialize(ValidDuration text) {
     }
 
-    public Event toAdminTriggerEvent(HttpServletRequest request, Authentication authentication, String clientUrl, String remoteAddr) {
-        var adminTrigger = new AdminTrigger();
-        adminTrigger.setRemoteAddr(remoteAddr);
-        adminTrigger.setTokenName(authentication.getName());
-        adminTrigger.setClientUrl(clientUrl);
-        return new Event(VERSION, adminTrigger);
+    @Override
+    public boolean isValid(String text, ConstraintValidatorContext cxt) {
+        try {
+            Duration.parse(text);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
     }
-
 }