+++ src/main/java/nl/dtls/fairdatapoint/entity/index/settings/IndexSettingsRetrieval.java
@@ -20,37 +20,30 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package nl.dtls.fairdatapoint.service.index.event;
+package nl.dtls.fairdatapoint.entity.index.settings;
 
+import lombok.*;
 
-import nl.dtls.fairdatapoint.api.dto.index.event.EventDTO;
-import nl.dtls.fairdatapoint.entity.index.event.AdminTrigger;
-import nl.dtls.fairdatapoint.entity.index.event.Event;
-import org.springframework.security.core.Authentication;
-import org.springframework.stereotype.Service;
+import javax.validation.constraints.NotNull;
+import java.time.Duration;
 
-import javax.servlet.http.HttpServletRequest;
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Builder(toBuilder = true)
+@EqualsAndHashCode
+public class IndexSettingsRetrieval {
+    @NotNull
+    private Duration rateLimitWait;
 
-@Service
-public class EventMapper {
+    @NotNull
+    private Duration timeout;
 
-    private static final Integer VERSION = 1;
-
-    public EventDTO toDTO(Event event) {
-        return new EventDTO(
-                event.getUuid(),
-                event.getType(),
-                event.getCreated().toString(),
-                event.getFinished().toString()
-        );
+    public static IndexSettingsRetrieval getDefault() {
+        IndexSettingsRetrieval retrieval = new IndexSettingsRetrieval();
+        retrieval.setRateLimitWait(Duration.ofMinutes(10));
+        retrieval.setTimeout(Duration.ofMinutes(1));
+        return retrieval;
     }
-
-    public Event toAdminTriggerEvent(HttpServletRequest request, Authentication authentication, String clientUrl, String remoteAddr) {
-        var adminTrigger = new AdminTrigger();
-        adminTrigger.setRemoteAddr(remoteAddr);
-        adminTrigger.setTokenName(authentication.getName());
-        adminTrigger.setClientUrl(clientUrl);
-        return new Event(VERSION, adminTrigger);
-    }
-
 }