+++ src/main/java/nl/dtls/fairdatapoint/util/HttpUtil.java
@@ -40,16 +40,43 @@ import static nl.dtls.fairdatapoint.util.ValueFactoryHelper.i;
 @Slf4j
 public class HttpUtil {
 
+    private static final String[] IP_HEADER_CANDIDATES = {
+            "X-Forwarded-For",
+            "X-Real-IP",
+            "Proxy-Client-IP",
+            "WL-Proxy-Client-IP",
+            "HTTP_X_FORWARDED_FOR",
+            "HTTP_X_FORWARDED",
+            "HTTP_X_CLUSTER_CLIENT_IP",
+            "HTTP_CLIENT_IP",
+            "HTTP_FORWARDED_FOR",
+            "HTTP_FORWARDED",
+            "HTTP_VIA",
+            "REMOTE_ADDR"
+    };
+
+    public static String getClientIpAddress(HttpServletRequest request, Boolean behindProxy) {
+        if (behindProxy) {
+            for (String header : IP_HEADER_CANDIDATES) {
+                String ipList = request.getHeader(header);
+                if (ipList != null && ipList.length() != 0 && !"unknown".equalsIgnoreCase(ipList)) {
+                    return ipList.split(",")[0];
+                }
+            }
+        }
+        return request.getRemoteAddr();
+    }
+
     public static String getRequestURL(HttpServletRequest request, String persistentUrl) {
         String urlS = request.getRequestURL().toString();
-        log.info("Original requesed url {}", urlS);
+        log.info("Original requested url {}", urlS);
         try {
             urlS = removeLastSlash(urlS.replace("/expanded", ""));
             persistentUrl = removeLastSlash(persistentUrl);
 
             URL url = new URL(urlS);
             String modifiedUrl = persistentUrl + url.getPath();
-            log.info("Modified requesed url {}", modifiedUrl);
+            log.info("Modified requested url {}", modifiedUrl);
 
             return modifiedUrl;