+++ src/test/java/nl/dtls/fairdatapoint/acceptance/index/ping/List_POST.java
@@ -43,7 +43,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 
-@DisplayName("POST /index")
+@DisplayName("POST /")
 public class List_POST extends WebIntegrationTest {
 
     @Autowired