+++ CHANGELOG.md
@@ -2,10 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
+to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.8.0]
+
+### Added
+
+- Denylist for FDP Index pings
+- Endpoints for managing Index settings from [Client]
+
+### Changed
+
+- Upgrade Java JDK from 14 to 15
+- Rate limits use forwarded IP by proxy based on config
+- Index settings are moved to the database
+- Admin trigger now accepts the same DTO as ping
+
+### Fixed
+
+- Fix metadata not found error
 
 ## [1.7.0]
 
@@ -155,4 +173,7 @@ The first release of reference FAIR Data Point implementation.
 [1.4.0]: /../../tree/v1.4.0
 [1.5.0]: /../../tree/v1.5.0
 [1.6.0]: /../../tree/v1.6.0
+
 [1.7.0]: /../../tree/v1.7.0
+
+[1.8.0]: /../../tree/v1.8.0